==> Comprehensive Machine Learning projects/Stock Portfolio Optimisation/2_JSE_Listed_Companies_Webscrapping.ipynb <==
import sys

import requests
from bs4 import BeautifulSoup
import re
import unicodedata
import pandas as pd
import numpy as np

url = 'https://sashares.co.za/shares-list/#gs.jxh54k'

# Get stock data from Sashares
data = requests.get(url).text

# Create a soup object 
soup = BeautifulSoup(data, 'html.parser')

# Find all the tables in the data
html_tables = soup.find_all('table')
# Count the tables
len(html_tables)

# We only want the first table containing stock data
stock_table = html_tables[0]

# Instantiate lists to store stock data
jse_code = []
share = []
company = []

# Get a list of rows, exclude table headings
table_rows = stock_table.find_all('tr')[1:]

# Check row data - what does it look like?
table_rows[0]

# Iterate over list of rows and extract the stock symbol, name and company name from each row
for i, row in enumerate(table_rows):
    cells = row.find_all('td')
    jse_code.append(cells[0].string)
    share.append(cells[1].string)
    company.append(cells[2].string)

# Create a DataFrame with the stock data
stock_data = pd.DataFrame({"JSE_code": jse_code, "Ticker": share, "Company": company})

# Check first few rows
stock_data.head()

# Check last few rows
stock_data.tail()

To clean the data we need to perform the following operations:
* Drop the column Ticker - it doesn't provide any useful information
* Drop the last row (527) which contains table headings
* Select stocks only - exclude Exchange Traded Notes
* Reset the index

# Drop column Ticker, not really useful.
stock_data.drop(columns=['Ticker'], inplace=True)
# Drop the last row containing heading data
stock_data.drop(index=527, inplace=True)
# Remove Exchange Traded Notes
stock_data = stock_data[stock_data['Company'] != 'Exchange Traded Notes']
# Remove Exchange Traded Funds
stock_data = stock_data[stock_data['Company'] != 'Exchange Traded Funds']
# Reset the index
stock_data.reset_index(inplace=True, drop=True)
# Now let's take a look
stock_data.iloc[40:60]

stock_data.shape

Next, we need to additional information about companies, such as market capitalisation and sector.

company_data = pd.read_csv('companies_list.csv')
# Displace the data
company_data.head()

# Create a new Dataframe to hold the combined info
stock_list = stock_data
# Add new empty columns for market capitalisation and sector
stock_list = stock_list.assign(Market_cap='', Sector='')

# Let's take a look at the new DataFrame
stock_list.head()

Next, we need match `JSE_code` in `stock_list` to `Code` in `company_data`. By matching the two, we can extract market cap and sector info for each available the stock from `company_data` and add it to `stock_list`. We'll try to do it in computationally less expensive way.

# Create lists of containing data we want to compare and extract.
jse_code_list = stock_list['JSE_code'].tolist()
comp_codes = company_data['Code'].tolist()
comp_marks = company_data['Market Cap'].tolist()
comp_sectors = company_data['Sector'].tolist()
# Create empty lists to store the matched stock ticker, extracted market cap and sector data
symbl_list=[]
mkt_list = []
sector_list = []
# Lets move down the list of stocks in stock_list - maintaining order is important when combining it later.
# For each stock symbol in stock list
for symbl in jse_code_list:
    # If the symbol is also in the stock list from company data.
    # using a set makes it faster and eliminates duplicates
    if symbl in set(comp_codes):
        # If the stock symbol matches, then
        for code, cap, sec in zip(comp_codes, comp_marks, comp_sectors):
            if code == symbl:
                # Add the stock symbol, market cap and sector to the empty lists
                symbl_list.append(code)
                mkt_list.append(cap)
                sector_list.append(sec)
    else:
        # If the stock is not in comp_codes
        # Still add the symbol to the list - maintain order and number
        # Add NAN for other missing values
        symbl_list.append(symbl)
        mkt_list.append(np.nan)
        sector_list.append(np.nan)

Let's verify our operation. The lengths of the lists must be sames as `stock_list`. 

# Check length of resulting lists
print(len(symbl_list))
print(len(mkt_list))
print(len(sector_list))
print(stock_list.shape)

Let's add the lists to `stock_list`

stock_list['Market_cap'] = mkt_list
stock_list['Sector'] = sector_list

# Check the results
stock_list.head(20)

Now let's save this as `stock_list.csv`. We will use this in the next stage of the project to select stock data.

stock_list.to_csv('stock_list.csv', index=False)
