==> Comprehensive Machine Learning projects/Stock Portfolio Optimisation/3_Fundamentals_data.ipynb <==
import yfinance as yf
import pandas as pd
import numpy as np
from collections import defaultdict

absa = yf.Ticker("ABG.JO")
stock_info = absa.info

fundamentals = ['symbol', 'shortName', 'sector', 'marketCap', 'trailingEps', 'trailingPE', 'returnOnEquity', 'priceToBook', 'payoutRatio',
                'priceToSalesTrailing12Months', 'freeCashflow', 'dividendYield', 'debtToEquity', 'pegRatio']

for attr in fundamentals:
    print(f'{attr} \t\t\t {stock_info[attr]}')

Import stock data

data = pd.read_csv('stock_list.csv')
# View first few rows
data.head()

Let's get the list of stocks. We need to add ".JO" to each ticker to signify that the stock data we want is for stocks listed on the JSE.

stock_list = data.JSE_code.tolist()
# Append ".JO"
stock_list = [x + ".JO" for x in stock_list]
# Check outcome
stock_list[0:5]

We need to get the latest market cap and sector information for each stock. In addition, we will also get the following fundamentals data:
* Earnings per share, EPS - trailing
* Price to earnings, PE ratio - trailing 
* Return on equity, ROE 
* Price to book value, PB ratio
* Dividend payout ratio
* Price to sales value, PS ratio - trailing
* Free Cash Flow, FCF
* Divident yield
* Debt to equity ratio,
* Projected earnings growth rate, PEG

We will store this information in lists which we will use to create a DataFrame.

# Create empty lists
list_names = ['ticker', 'name', 'sector', 'market_cap', 'EPS', 'PE_ratio', 'ROE', 'PB_ratio', 
              'div_payout_ratio', 'PS_ratio', 'FCF', 'divident_yield', 'debt_equity', 'earnings_growth']

# Create a dictionary to store our lists
stock_fund = defaultdict(list)

Let's get the fundamentals data from Yahoo! Finance. This process will take a while...

# For every stock in our list.
for ticker in stock_list:
    # get stock data and store it in 'data'
    data = yf.Ticker(ticker)
    # we need the stock info, store it in 'info'
    info = data.info
    # let's get the fundamentals data and store it in a list with the same name
    for lst, attr in zip(list_names, fundamentals):
        try:
            stock_fund[lst].append(info[attr])  
        except KeyError: 
            stock_fund[lst].append(np.nan)

print('Processing done')

Let's check the length of each of list

for name, array in stock_fund.items():
    print(f'{name} \t {len(array)}')


Now that we have all the stock info we need, we'll store it in a DataFrame. 

# Create DataFrame
df = pd.DataFrame(stock_fund)
# Check few rows
df.head()

Let's save the data as a csv file `fundamentals.csv`. In the next stage of the process, we'll clean the data.

df.to_csv('fundamentals.csv', index=False)
