==> Comprehensive Machine Learning projects/Stock Portfolio Optimisation/4_Data_Wrangling.ipynb <==
# **Portfolio Optimisation and Backtesting**

## Data wrangling

In this part of the project, we will focus on cleaning the fundamentals data we obtained from Yahoo! Finance.

This step will prepare the data for stock screening, one of the key steps in the project.

## Objectives

In this noebook we will:

*   Handle missing values
*   Format values

<h2>Table of Contents</h2>

<div class="alert alert-block alert-info" style="margin-top: 20px">
<ul>
    <li><a href="#identify_handle_missing_values">Identify and handle missing values</a>
        <ul>
            <li><a href="#identify_missing_values">Identify missing values</a></li>
            <li><a href="#deal_missing_values">Deal with missing values</a></li>
            <li><a href="#correct_data_format">Format currency values</a></li>
        </ul>
    </li>
</ul>

</div>

<hr>

Let's start by importing the necessary libraries.

import pandas as pd
import numpy as np
import matplotlib as mpl
import matplotlib.pyplot as plt
%matplotlib inline

Next, we'll import the companies'fundamentals data from `fundamentals.csv`.

df = pd.read_csv('fundamentals.csv')
# Check the first few rows
df.head()

Its important that we have an idea of our data before we begin the cleaning process. Let's start with a general info of the dataframe.

df.info()

df.shape

Already we can see that some of the columns have NaN and zero values which make it hard to analyse our data. 

We intend to screen our stocks using these different attributes, so we'll have to drop columns where there's a lot of missing data (NaNs and zeros). Replacing these values with either the industry median or mean might not be appropriate because these values are the basis of screening similar stocks (in the same sector) when building a diversified portfolio. 

df.columns

We'll capitalise the beginning of each column name for easier referencing and also correct 'divident'. 

df.rename(columns={'ticker':'Ticker', 'name':'Name', 'sector':'Sector', 'market_cap':'Market_cap',
                   'div_payout_ratio':'Div_payout', 'divident_yield':'Div_yield', 'debt_equity':'Debt_to_equity', 
                   'earnings_growth':'Earnings_growth'}, inplace=True)
df.columns

## Missing values

We can deal with missing values in three way:
* Look for these values elsewhere - which might not be feasible
* Drop missing values or complete columns if a significant number of values are missing.
* Fill missing values with either median or average values

Let's look at the number of missing values in each column again.

df.isnull().mean().round(2)*100

df.head(10)

We'll drop every row where `Ticker` is NaN.

# Avoid changing original data 
df_edited = df.copy()
# Drop rows with missing values in 'Ticker' and 'Name'.
df_edited.dropna(axis=0, subset=['Ticker', 'Name'], inplace=True)

About 85% of Earnings growth rate numbers are missing from our data. Its difficult to infer values for growth from other variables or sources, so will drop the entire column. Since we have Dividend payout with only 16% of values missing, we can drop Dividend yield which is missing about 43% of values. Although they are different, they're still related and payout ratio seems to be the better indicator when forward looking. 

# Drop columns
df_edited.drop(columns=['Earnings_growth', 'Div_yield'], axis=1, inplace=True)

In Sector, missing values represent REITS, Funds and preference shares mostly. There are less than 10 entries for these. Since we will group values by sector and fill in other missing values based on sector, we will drop missing values from this column.

# Drop rows with missing values in 'Sector'.
df_edited.dropna(axis=0, subset=['Sector'], inplace=True)

Now, lets create a matrix of average values for each metric by sector. We will use these values to replace `NaN` values. Using sector averages for a metric, like EPS, makes more economic sense than using the mean EPS that takes into account all companies since the economic environment is different from one sector to another. It is common practice in investment analysis to compare a company to its peers. 

averages = df_edited.groupby(by='Sector').mean()
averages

We have a missing `PS_ratio` value for Financial Services. We'll replace it with the median value instead.

median = df_edited[['Sector', 'PS_ratio']].groupby(by='Sector').median().loc['Financial Services', 'PS_ratio']

averages.loc['Financial Services', 'PS_ratio'] = median

Next, we'll iterate over each row of the data and check whether there are any missing values for a particular stock. If a value is missing, we'll replace it with the average value from `averages`. Because its unadvisable to change values were are iterating over, we'll store the return values in lists.

# Create empty lists to store the return values
Ticker = []
Name = []
Sector = []
Market_cap = []
EPS = []
PE_ratio = []
ROE = []
PB_ratio = []
Div_payout = []
PS_ratio = []
FCF = []
Debt_to_equity = []

# For each row (stock) in df_edited
# Ignore index value and get row values only
for _, row in df_edited.iterrows():
    # Return ticker, name and sector as they are
    Ticker.append(row.Ticker)
    Name.append(row.Name)
    Sector.append(row.Sector)
    # Get the stock's sector
    sector = row.Sector
    # Use the sector value and Market_cap to look up average value in averages
    # Add the look up value from averages to return list If the original value is NaN
    # Else (if original value is not NaN), return the original value as it is
    Market_cap.append(averages.loc[sector, 'Market_cap'] if np.isnan(row.Market_cap) else row.Market_cap)
    # Repeat the same for all other metric (column) values
    EPS.append(averages.loc[sector, 'EPS'] if np.isnan(row.EPS) else row.EPS)
    PE_ratio.append(averages.loc[sector, 'PE_ratio'] if np.isnan(row.PE_ratio) else row.PE_ratio)
    ROE.append(averages.loc[sector, 'ROE'] if np.isnan(row.ROE) else row.ROE)
    PB_ratio.append(averages.loc[sector, 'PB_ratio'] if np.isnan(row.PB_ratio) else row.PB_ratio) 
    Div_payout.append(averages.loc[sector, 'Div_payout'] if np.isnan(row.Div_payout) else row.Div_payout)  
    PS_ratio.append(averages.loc[sector, 'PS_ratio'] if np.isnan(row.PS_ratio) else row.PS_ratio) 
    FCF.append(averages.loc[sector, 'FCF'] if np.isnan(row.FCF) else row.FCF) 
    Debt_to_equity.append(averages.loc[sector, 'Debt_to_equity'] if np.isnan(row.Debt_to_equity) else row.Debt_to_equity)

We'll create a new DataFrame `df_filled` using the lists from above.

df_filled = pd.DataFrame(data={'Ticker':Ticker, 'Name':Name, 'Sector':Sector, 'Market_cap':Market_cap, 'EPS':EPS, 'PE_ratio':PE_ratio, 'ROE':ROE, 'PB_ratio':PB_ratio, 
                             'Div_payout':Div_payout, 'PS_ratio':PS_ratio, 'FCF':FCF, 'Debt_to_equity':Debt_to_equity})

Now let's view the outcome of our operation by comparing `df_edited` to `df_filled`.  

df_edited.head()

df_filled.head()

Lets check our data for missing values again.

df_filled.isnull().mean().round(2)*100

df_filled.shape

## Format values

Currency values for market cap, earnings per share and free cash flow are quoted by Yahoo! Finance in cents. We'll convert Market cap and Free cash flow values to Rand amounts to make it intuitve. We'll leave EPS as it is. 

# Divide all Market_cap values by 100 to convert them to Rands
df_filled.Market_cap = df_filled.Market_cap.apply(lambda x: x/100)
# Do the same for FCF
df_filled.FCF = df_filled.FCF.apply(lambda x: x/100)
# Verify result
df_filled.head()

It worked. Prior, market cap for ABSA Bank was ZAc 363,123,800,000 (3.631238e+11). Now its ZAR 3,631,238,000 (3.631238e+09) - roughly R3.6 billion.

Now that our data is clean, we can start analysing the data and selecting the stocks we want in our portfolio in the next stage of the project. We'll save this data as 'jse_stock_data_cleaned.csv'

df_filled.to_csv('jse_stock_data_cleaned.csv', index=False)
